a bit of cleaning up. removing unused code.
